・Dao用にconfigクラスを分けた ・健康情報履歴画面でヘッダーに単位追加
