+++ app-manage/src/main/java/jp/co/isol/manage/config/ManageConfig.java
@@ -3,12 +3,6 @@ package jp.co.isol.manage.config;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
-import jp.co.isol.common.dao.AccountDao;
-import jp.co.isol.common.dao.HealthInfoDao;
-import jp.co.isol.common.dao.MailInfoDao;
-import jp.co.isol.common.dao.impl.AccountDaoImpl;
-import jp.co.isol.common.dao.impl.HealthInfoDaoImpl;
-import jp.co.isol.common.dao.impl.MailInfoDaoImpl;
 import jp.co.isol.common.other.DBConnecter;
 import jp.co.isol.manage.log.ManageLogger;
 import jp.co.isol.manage.web.session.ManageSessionManager;
@@ -34,22 +28,6 @@ public class ManageConfig {
 	@Bean(name = "manageLogger")
 	public ManageLogger getManageLogger() {
 		return new ManageLogger();
-	}
-
-	/* Dao */
-	@Bean(name = "accountDao")
-	public AccountDao getAccountDao() {
-		return new AccountDaoImpl();
-	}
-
-	@Bean(name = "healthInfoDao")
-	public HealthInfoDao getHealthInfoDao() {
-		return new HealthInfoDaoImpl();
-	}
-
-	@Bean(name = "mailInfoDao")
-	public MailInfoDao getMailInfoDao() {
-		return new MailInfoDaoImpl();
 	}
 
 }
